Move all integration tests into one test binary
- Fewer binaries to link for testing
- Allows running specific integration test suite, e.g. `cargo test sqs`
- Allows running more tests in parallel

See also:
https://matklad.github.io/2021/02/27/delete-cargo-integration-tests.html
