+++ omniqueue/tests/it/main.rs
@@ -1,0 +1,10 @@
+#[cfg(feature = "gcp_pubsub")]
+mod gcp_pubsub;
+#[cfg(feature = "rabbitmq")]
+mod rabbitmq;
+#[cfg(feature = "redis")]
+mod redis;
+#[cfg(feature = "redis_cluster")]
+mod redis_cluster;
+#[cfg(feature = "sqs")]
+mod sqs;
