+++ omniqueue/tests/it/redis.rs
@@ -1,5 +1,3 @@
-#![cfg(feature = "redis")]
-
 use omniqueue::{
     backends::redis::{RedisConfig, RedisQueueBackend},
     queue::{consumer::QueueConsumer, producer::QueueProducer, QueueBackend, QueueBuilder, Static},
