+++ omniqueue/tests/it/sqs.rs
@@ -1,5 +1,3 @@
-#![cfg(feature = "sqs")]
-
 use aws_sdk_sqs::Client;
 use omniqueue::{
     backends::sqs::{SqsConfig, SqsQueueBackend},
